requirements.txt
pandas
numpy
scikit-learn
xgboost
matplotlib
plotly
shap
nltk
spacy
bertopic
umap-learn
pyyaml
